Initial commit
This library was extracted from my Arduino UNO R4 Test
github projectL https://github.com/KurtE/UNOR4-stuff

However I renamed the library and class to fit with the Arduino Library requirements, and fixed some of the issues that Had.

Just doing this for the fun of it, so no promises.  but thought some others might have fun with it as well.